+++ src/main/java/mcjty/incontrol/rules/SpecialRule.java
@@ -37,12 +37,12 @@ public class SpecialRule extends RuleBase<RuleBase.EventGetter> {
 
         @Override
         public BlockPos getPos(LivingSpawnEvent.SpecialSpawn o) {
-            return new BlockPos(o.getX(), o.getY(), o.getZ());
+            return new BlockPos((int) o.getX(), (int) o.getY(), (int) o.getZ());
         }
 
         @Override
         public BlockPos getValidBlockPos(LivingSpawnEvent.SpecialSpawn o) {
-            return new BlockPos(o.getX(), o.getY() - 1, o.getZ());
+            return new BlockPos((int) o.getX(), (int) (o.getY() - 1), (int) o.getZ());
         }
 
         @Override
@@ -67,7 +67,7 @@ public class SpecialRule extends RuleBase<RuleBase.EventGetter> {
 
         @Override
         public Player getPlayer(LivingSpawnEvent.SpecialSpawn o) {
-            return getClosestPlayer(o.getLevel(), new BlockPos(o.getX(), o.getY(), o.getZ()));
+            return getClosestPlayer(o.getLevel(), new BlockPos((int) o.getX(), (int) o.getY(), (int) o.getZ()));
         }
 
         @Override
